fix(bicep): resolve retryPolicy deserialization error
Moved the 'retryPolicy' block inside the 'inputs' block for the 'Get_Oauth_Token' and 'Get_Machines' Http actions.

This change is required to address a deserialization error during Logic App deployment, where the 'retryPolicy' is not a recognized member of the 'FlowTemplateAction' object at the top level of the action definition. Placing it within the 'inputs' block aligns with the expected schema for this specific context.